post eia api v2 deprecations
* Deleted old cleaning functions and renamed clean_v2 to clean

* Warning now issued when instantiating BaData
